feat: add future class to manage async GPU computations (PROOF-171) (#155)
* add future

* support implicit stream conversions

* fix ownership issues

* tweak move assignment operator

* fix memory ownership in test code

+++ sxt/execution/async/test_kernel.cc
@@ -7,7 +7,7 @@ namespace sxt::xena {
 //--------------------------------------------------------------------------------------------------
 // add_impl
 //--------------------------------------------------------------------------------------------------
-static __global__ void add_impl(double* c, double* a, double* b, int n) {
+static __global__ void add_impl(uint64_t* c, uint64_t* a, uint64_t* b, int n) {
   int i = blockIdx.x * blockDim.x + threadIdx.x;
   if (i < n) {
     c[i] = a[i] + b[i];
@@ -17,7 +17,8 @@ static __global__ void add_impl(double* c, double* a, double* b, int n) {
 //--------------------------------------------------------------------------------------------------
 // add_for_testing
 //--------------------------------------------------------------------------------------------------
-void add_for_testing(double* c, const xenb::stream& stream, double* a, double* b, int n) noexcept {
+void add_for_testing(uint64_t* c, const xenb::stream& stream, uint64_t* a, uint64_t* b,
+                     int n) noexcept {
   add_impl<<<basn::divide_up(n, 256), 256, 0, stream.raw_stream()>>>(c, a, b, n);
 }
 } // namespace sxt::xena
